third_party/WebKit/LayoutTests/images/image-use-counters.html

<!DOCTYPE html>
<title>ImageDocument UseCounters</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<iframe></iframe>
<script>
async_test(function() {
  // From UseCounter.h
  var ImageDocument = 1274;
  var ImageDocumentInFrame = 1310;
  assert_false(internals.isUseCounted(document, ImageDocument));
  assert_false(internals.isUseCounted(document, ImageDocumentInFrame));
  var iframe = document.querySelector("iframe");
  iframe.src = "resources/dice.png";
  iframe.onload = this.step_func_done(function() {
    assert_true(internals.isUseCounted(document, ImageDocument));
    assert_true(internals.isUseCounted(document, ImageDocumentInFrame));
  });
});
</script>
